Refactor: payment 부분 리펙토링
1. payment 파일 삭제
2. paymentTest -> payment
3. 결제 연결 주문 코드 변경
