+++ src/main/java/com/shineidle/tripf/order/entity/Order.java
@@ -2,8 +2,8 @@ package com.shineidle.tripf.order.entity;
 
 import com.shineidle.tripf.common.entity.BaseEntity;
 import com.shineidle.tripf.order.type.OrderStatus;
+import com.shineidle.tripf.order.type.PayStatus;
 import com.shineidle.tripf.orderProduct.entity.OrderProduct;
-import com.shineidle.tripf.payment.entity.Payment;
 import com.shineidle.tripf.product.entity.Product;
 import com.shineidle.tripf.user.entity.User;
 import jakarta.persistence.*;
@@ -28,6 +28,9 @@ public class Order extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus = OrderStatus.ORDER_RECEIVED;
 
+    @Enumerated(EnumType.STRING)
+    private PayStatus payStatus = PayStatus.PENDING;
+
     @ManyToOne
     private Product product;
 
@@ -37,24 +40,19 @@ public class Order extends BaseEntity {
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<OrderProduct> orderProducts = new ArrayList<>();
 
-    @OneToOne(mappedBy = "order", cascade = CascadeType.ALL)
-    private Payment payment;
-
     private LocalDateTime createdAt;
 
     public Order() {
     }
 
-    public Order(Long totalPrice, OrderStatus orderStatus, Product product, User user, LocalDateTime createdAt, List<OrderProduct> orderProducts, Payment payment) {
+    public Order(Long totalPrice, OrderStatus orderStatus, PayStatus payStatus, Product product, User user, LocalDateTime createdAt) {
         this.totalPrice = totalPrice;
         this.orderStatus = orderStatus;
+        this.payStatus = payStatus;
         this.product = product;
         this.user = user;
         this.createdAt = createdAt;
-        this.orderProducts = orderProducts != null ? orderProducts : new ArrayList<>();
-        this.payment = payment;
     }
-
 
     public void addOrderProduct(OrderProduct orderProduct) {
         this.orderProducts.add(orderProduct);
@@ -66,10 +64,5 @@ public class Order extends BaseEntity {
         this.totalPrice = this.orderProducts.stream()
                 .mapToLong(OrderProduct::getTotalPrice)
                 .sum();
-    }
-
-    // 결제 설정 메서드 추가
-    public void setPayment(Payment payment) {
-        this.payment = payment;
     }
 }
