+++ src/main/java/com/shineidle/tripf/order/type/PayStatus.java
@@ -1,0 +1,7 @@
+package com.shineidle.tripf.order.type;
+
+public enum PayStatus {
+    PENDING, // 결제대기
+    PAID, //결제완료
+    FAILED //결제실패
+}
