--- src/main/java/com/shineidle/tripf/payment/dto/ApiResponse.java
@@ -1,22 +1,0 @@
-package com.shineidle.tripf.payment.dto;
-
-public class ApiResponse {
-    private String code;
-    private String message;
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
