--- src/main/java/com/shineidle/tripf/payment/type/PaymentStatus.java
@@ -1,7 +1,0 @@
-package com.shineidle.tripf.payment.type;
-
-public enum PaymentStatus {
-    PENDING, // 결제대기
-    PAID, //결제완료
-    FAILED //결제실채
-}
