--- src/main/java/com/shineidle/tripf/paymentTest/dto/PaymentRequestDtoTest.java
@@ -1,12 +1,0 @@
-package com.shineidle.tripf.paymentTest.dto;
-
-import lombok.Getter;
-
-@Getter
-public class PaymentRequestDtoTest {
-    private int amount;
-
-    public int getAmount() {
-        return amount;
-    }
-}
