--- src/main/java/com/shineidle/tripf/paymentTest/repository/PaymentRepositoryTest.java
@@ -1,8 +1,0 @@
-package com.shineidle.tripf.paymentTest.repository;
-
-import com.shineidle.tripf.paymentTest.entity.PaymentTest;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PaymentRepositoryTest extends JpaRepository<PaymentTest, Long> {
-    PaymentTest findByOrderId(String orderId);
-}
