--- src/main/resources/templates/payment/payConfirmation.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>결제 승인</title>
-</head>
-<body>
-<h1>결제 승인 결과</h1>
-
-<div>
-    <p>결제 성공! </p>
-    <p>결제 Key: <span th:text="${payment.paymentKey}"></span></p>
-    <p>결제 금액: <span th:text="${payment.amount}"></span></p>
-    <p>결제 상태:
-        <span th:text="${payment.cancelStatus == 'SUCCESS' ? '성공' : (payment.cancelStatus == 'FAILED' ? '실패' : '대기')}"></span>
-    </p>
-</div>
-
-<div>
-    <a href="/payments/details/{paymentKey}" th:href="@{/payments/details/{paymentKey}(paymentKey=${payment.paymentKey})}">결제 상세보기</a>
-</div>
-</body>
-</html>
