--- src/main/resources/templates/payment/payDetails.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>결제 내역</title>
-</head>
-<body>
-<h1>결제 내역</h1>
-
-<div>
-  <p><strong>결제 Key:</strong> <span th:text="${payment.paymentKey}"></span></p>
-  <p><strong>결제 URL:</strong> <span th:text="${payment.paymentUrl}"></span></p>
-  <p><strong>결제 금액:</strong> <span th:text="${payment.amount}"></span></p>
-  <p><strong>주문 ID:</strong> <span th:text="${payment.orderId}"></span></p>
-  <p><strong>사용자 ID:</strong> <span th:text="${payment.userId}"></span></p>
-  <p><strong>결제 상태:</strong>
-    <span th:text="${payment.cancelStatus == 'SUCCESS' ? '성공' : (payment.cancelStatus == 'FAILED' ? '실패' : '대기')}"></span>
-  </p>
-  <p><strong>결제 승인 일시:</strong> <span th:text="${#dates.format(payment.approvedAt, 'yyyy-MM-dd HH:mm:ss')}"></span></p>
-  <p><strong>결제 취소 일시:</strong> <span th:text="${#dates.format(payment.canceledAt, 'yyyy-MM-dd HH:mm:ss')}"></span></p>
-</div>
-
-<div>
-  <a href="/payments/cancel/details/{paymentKey}" th:href="@{/payments/cancel/details/{paymentKey}(paymentKey=${payment.paymentKey})}">결제 취소 상세 보기</a>
-</div>
-</body>
-</html>
