--- src/main/resources/templates/payment/payRequest.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>결제 요청</title>
-  <script src="https://js.tosspayments.com/v1/payment"></script>
-</head>
-<body>
-<h1>결제하기</h1>
-<p>주문 ID: <span th:text="${paymentRequest.orderId}"></span></p>
-<p>상품명: <span th:text="${paymentRequest.orderName}"></span></p>
-<p>금액: <span th:text="${paymentRequest.amount}"></span> 원</p>
-<p>고객 이름: <span th:text="${paymentRequest.customerName}"></span></p>
-
-<button id="paymentButton">결제하기</button>
-
-<script>
-  var clientKey = "test_ck_DLJOpm5QrlDZnZeKMBMLrPNdxbWn";  // 토스 결제용 클라이언트 키
-  var tossPayments = TossPayments(clientKey);
-
-  document.getElementById('paymentButton').addEventListener('click', function() {
-    // 결제 요청 데이터를 서버에서 받아오기
-    fetch('/payments/create', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        orderId: /*[[${paymentRequest.orderId}]]*/,  // 타임리프에서 전달된 값
-        orderName: /*[[${paymentRequest.orderName}]]*/,
-        amount: /*[[${paymentRequest.amount}]]*/,
-        customerName: /*[[${paymentRequest.customerName}]]*/
-      })
-    })
-            .then(response => response.json())
-            .then(data => {
-              // 서버에서 받은 결제 정보를 토대로 결제창 띄우기
-              tossPayments.requestPayment("카드", {
-                amount: data.amount,  // 백엔드에서 받은 결제 금액
-                orderId: data.orderId,  // 백엔드에서 받은 주문 ID
-                orderName: data.orderName,  // 백엔드에서 받은 상품명
-                customerName: data.customerName,  // 고객 이름
-                successUrl: "/payments/success",  // 결제 성공 시 이동할 URL
-                failUrl: "/payments/fail"  // 결제 실패 시 이동할 URL
-              })
-                      .catch(function(error) {
-                        console.error("결제 실패: ", error);
-                        alert("결제에 실패했습니다!");
-                      });
-            })
-            .catch(error => {
-              console.error('결제 요청 중 오류 발생:', error);
-              alert("결제 요청 중 오류가 발생했습니다.");
-            });
-  });
-</script>
-</body>
-</html>
